isolate stencil_tests grid fixture
