+++ model/atmosphere/diffusion/tests/diffusion/stencil_tests/conftest.py
@@ -1,0 +1,10 @@
+# ICON4Py - ICON inspired code in Python and GT4Py
+#
+# Copyright (c) 2022-2024, ETH Zurich and MeteoSwiss
+# All rights reserved.
+#
+# Please, refer to the LICENSE file in the root directory.
+# SPDX-License-Identifier: BSD-3-Clause
+
+from icon4py.model.testing.fixtures.stencil_tests import grid
+from icon4py.model.testing.fixtures.datatest import backend
